# docs/v1/swagger.yaml
definitions:
  forms.SigninRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  forms.SigninResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      session_id:
        type: string
    type: object
  forms.SignupRequest:
    properties:
      email:
        maxLength: 120
        type: string
      name:
        maxLength: 120
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  forms.SignupResponse:
    properties:
      email:
        maxLength: 120
        type: string
      name:
        maxLength: 120
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  forms.WhoAmIResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        $ref: '#/definitions/utils.ResponseStatus'
    type: object
  utils.ResponseStatus:
    enum:
    - success
    - error
    type: string
    x-enum-varnames:
    - ResponseStatusSuccess
    - ResponseStatusError
info:
  contact:
    email: dimasbagussusilo@gmail.com
    name: Dimas Bagus Susilo
    url: http://www.linkedin.com/in/dimasbagussusilo
  description: This server provides the Filesystem API needs.
  termsOfService: http://swagger.io/terms/
  title: Filesystem API
  version: "1.0"
paths:
  /api/v1/auth/signin:
    post:
      consumes:
      - application/json
      description: login user with credentials.
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/forms.SigninRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/forms.SigninResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Login user.
      tags:
      - Auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: register user.
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/forms.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/forms.SignupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Signup user.
      tags:
      - Auth
  /api/v1/users:
    get:
      consumes:
      - '*/*'
      description: get all users' data.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: Email
        in: query
        name: email
        type: string
      - description: Order by
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Show all user.
      tags:
      - Users
  /api/v1/users/me:
    get:
      consumes:
      - '*/*'
      description: get logged-in user data.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/forms.WhoAmIResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Show logged-in user.
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"